fix warnings

+++ src/skygfx/skygfx.cpp
@@ -386,12 +386,12 @@ uint32_t skygfx::GetMipCount(uint32_t width, uint32_t height)
 
 uint32_t skygfx::GetMipWidth(uint32_t base_width, uint32_t mip_level)
 {
-	return glm::max<uint32_t>(1, glm::floor<uint32_t>(base_width >> mip_level));
+	return glm::max<uint32_t>(1, static_cast<uint32_t>(glm::floor<uint32_t>(base_width >> mip_level)));
 }
 
 uint32_t skygfx::GetMipHeight(uint32_t base_height, uint32_t mip_level)
 {
-	return glm::max<uint32_t>(1, glm::floor<uint32_t>(base_height >> mip_level));
+	return glm::max<uint32_t>(1, static_cast<uint32_t>(glm::floor<uint32_t>(base_height >> mip_level)));
 }
 
 // device
